feat: create modules, new switcher, add eslint
