+++ style.css
@@ -1,3 +1,7 @@
+* {
+    box-sizing: border-box;
+}
+
 html {
     font-family: 'Roboto', sans-serif;
 }
@@ -13,13 +17,53 @@ html {
     justify-content: space-between;
 }
 
+.back-to-main {
+    margin-top: 1vw;
+    outline: none;
+    user-select: none;
+    border: none;
+    text-align: center;
+    height: max-content;
+    background: rgb(93,173,226);
+    background: linear-gradient(180deg, rgba(93,173,226,1) 0%, rgba(72,201,176,1) 100%, rgba(255,254,253,1) 100%); 
+    border-radius: 5px;
+    cursor: pointer;
+    transition: .2s;
+}
+
+.back-to-main-active {
+    background: rgb(245,176,65);
+    background: linear-gradient(180deg, rgba(245,176,65,1) 0%, rgba(236,112,99,1) 100%, rgba(255,254,253,1) 100%);
+}
+
+.back-to-main:active {
+    opacity: 0.5;
+}
+
+.main-link {
+    display: inline-flex;
+    width: 7vw;
+    height: 2vw;
+    padding: 1vw;
+    align-items: center;
+    color: #fff;
+    text-decoration: none;
+    font-size: 1vw;
+    transition: .2s;
+}
+
+.main-link:hover {
+    color: rgb(245,176,65);
+}
+
 .menu-switcher {
     display: flex;
     flex-direction: column;
     height: 26px;
     justify-content: space-between;
     cursor: pointer;
-    z-index: 2;
+    z-index: 4;
+    margin-top: 1vw;
 }
 
 .menu-switcher:active {
@@ -52,10 +96,10 @@ html {
 
 input[type="checkbox"]{
     position: relative;
-    width: 120px;
-    height: 40px;
+    width: 122px;
+    height: 42px;
     background: rgb(93,173,226);
-    background: linear-gradient(180deg, rgba(93,173,226,1) 0%, rgba(72,201,176,1) 51%, rgba(255,254,253,1) 100%); 
+    background: linear-gradient(180deg, rgba(93,173,226,1) 0%, rgba(72,201,176,1) 100%, rgba(255,254,253,1) 100%); 
     -webkit-appearance: none;
     border-radius: 5px;
     outline: none;
@@ -66,19 +110,15 @@ input[type="checkbox"]{
 
 input:checked[type="checkbox"]{
     background: rgb(245,176,65);
-    background: linear-gradient(180deg, rgba(245,176,65,1) 0%, rgba(236,112,99,1) 51%, rgba(255,254,253,1) 100%);
+    background: linear-gradient(180deg, rgba(245,176,65,1) 0%, rgba(236,112,99,1) 100%, rgba(255,254,253,1) 100%);
 }
 
 input[type="checkbox"]::before{
     z-index: 2;
     position: absolute;
-    content: '';
+    content: url('assets/images/scroll.png');
     left: 2px;
-    top: 2px;
-    width: 40px;
-    height: 36px;
-    background: #fff;
-    border-radius: 3px;
+    top: 1px;
     transition: .4s;
 }
 
@@ -178,47 +218,19 @@ label {
     justify-content: space-around;
     align-items: center;
     background: -webkit-gradient(linear,left top,left bottom,from(#5DADE2 ),color-stop(40%,#48C9B0),color-stop(40%,#fff),to(#fff));
-    background-color: linear-gradient(240deg,#5DADE2,#48C9B0 40%,#fff 0,#fff);
     text-decoration: none;
     font-size: 1.5vw;
     color: black;
     box-shadow: 0 8px 6px -6px #2E86C1  ;
     transition: .2s ease-in-out;
-    perspective: 1000px;
+    -webkit-perspective: 1000;
+    position: relative;
 }
 
 .category-card-play {
     background: -webkit-gradient(linear,left top,left bottom,from(#F5B041   ),color-stop(40%,#EC7063 ),color-stop(40%,#fff),to(#fff));
     background-color: linear-gradient(240deg,#F5B041  ,#EC7063  40%,#fff 0,#fff);
     box-shadow: 0 8px 6px -6px #F5B041  ;
-}
-
-.category-inner {
-    position: relative;
-    width: 100%;
-    height: 100%;
-    transition: transform 0.8s;
-    transform-style: preserve-3d;
-}
-
-.hover {
-    transform: rotateY(180deg);
-}
-
-.category-front, .category-back {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    backface-visibility: hidden;
-}
-
-.category-back {
-    transform: rotateY(180deg);
-    text-align: center;
-}
-
-.category-card-transfom {
-    transform: rotateY(180deg);
 }
 
 .category-card:hover {
@@ -234,15 +246,60 @@ label {
     transform: scale(0.95);
 }
 
-.category-word {
+.category-word-container {
+    margin: 2vw;
+    border-radius: 5px;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-around;
+    align-items: center;
+    font-size: 1.5vw;
+    color: black;
+    box-shadow: 0 8px 6px -6px #2E86C1  ;
+    transition: .2s ease-in-out;
+    -webkit-perspective: 800;
+    position: relative;
     cursor: pointer;
     background: #fff;
     height: 13vw;
     width: 15vw;
 }
 
-.category-word:hover {
-    transform: none;
+.transparent {
+    background: transparent;
+    box-shadow: none;
+}
+
+.category-word-container .category-inner.flipped {
+    -webkit-transform: rotateY(-180deg);
+}
+
+.category-word-container .category-inner {
+    width: 100%;
+    height: 100%;
+    -webkit-transform-style: preserve-3d;
+    -webkit-transition: 0.8s;
+}
+
+.category-word-container .category-inner .face {
+    width: 100%;
+    height: 100%;
+    position: absolute;
+    -webkit-backface-visibility: hidden;
+    z-index: 2;
+}
+
+.category-word-container .category-inner .category-front {
+    position: absolute;
+    z-index: 1;
+}
+
+.category-word-container .category-inner .category-back {
+    -webkit-transform: rotateY(-180deg);
+    text-align: center;
+}
+
+.category-word-container:hover {
     filter: drop-shadow(0 0 1vw #2E86C1 );
 }
 
@@ -331,18 +388,19 @@ label {
 }
 
 .stars-line {
-    position: absolute;
     display: flex;
     flex-direction: row;
     flex-wrap: nowrap;
-    top: -5%;
-    right: 0;
     direction: rtl;
     overflow-x: hidden;
+    width: 94%;
+    margin: 0 auto;
+    height: 40px;
 }
 
 .star {
-    width: 70%;
+    width: 100%;
+    max-width: 40px;
     height: auto;
 }
 
@@ -351,7 +409,39 @@ label {
     list-style: none;
     display: none;
 }
+
+.success {
+    position: fixed;
+    top: 0;
+    left: 0;
+    z-index: 4;
+    background: #fff;
+    background-image: url('assets/images/success.jpg');
+    background-repeat: no-repeat;
+    background-position: center;
+    text-align: center;
+    font-size: 30px;
+    color: brown;
+    width: 100%;
+    height: 100vh;
+}
+
+.success p {
+    margin-top: 10%;
+}
+
+.failure {
+    background-image: url('assets/images/failure.jpg'); 
+}
 @media (max-width: 768px) {
+    .back-to-main {
+        margin-top: 0.5vw;
+    }
+    .main-link {
+        width: 15vw;
+        height: 5vw;
+        font-size: 2.6vw;
+    }
     .sliding-menu {
         width: 40vw;
     }
@@ -364,10 +454,61 @@ label {
         height: 35vw;
         font-size: 3.5vw;
     }
-    .category-word {
+    .category-word-container {
+        margin: 2vw;
+        width: 35vw;
         height: 30vw;
+        font-size: 3.5vw;
     }
     .card-playing {
         height: 23vw;
+    }
+}
+@media (max-width: 500px) {
+    .menu-switcher {
+        margin-top: 3vw;
+    }
+    .main-link {
+        width: 20vw;
+        height: 11vw;
+        font-size: 4vw;
+    }
+    .sliding-menu {
+        width: 65vw;
+        z-index: 3;
+    }
+    .category-link {
+        font-size: 6vw;
+        line-height: 3;
+    }
+    .category-card {
+        margin-bottom: 10vw;
+        width: 80vw;
+        height: 70vw;
+        font-size: 3.5vw;
+        font-size: 7vw;
+    }
+    .category-word-container {
+        margin-bottom: 10vw;
+        width: 80vw;
+        height: 63vw;
+        font-size: 3.5vw;
+        font-size: 7vw;
+    }
+    .success {
+        background-size: contain;
+    }
+    .card-playing {
+        height: 48vw;
+    }
+    .play-button {
+        height: 16vw;
+        font-size: 10vw;
+    }
+}
+@media (max-width: 320px) {
+    .play-button {
+        height: 19vw;
+        font-size: 10vw;
     }
 }
